update config add vm func and deploy func ,include Fingerprint and priivate network config
